activate hardhat provider at load
