+++ libs/remix-ui/run-tab/src/lib/run-tab.tsx
@@ -61,6 +61,7 @@ export function RunTabUI (props: RunTabProps) {
 
   useEffect(() => {
     initRunTab(plugin)(dispatch)
+    plugin.onInitDone()
   }, [plugin])
 
   useEffect(() => {
